Fixes reading profile error when running on WASM
